+++ CHANGELOG.md
@@ -8,12 +8,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 You can find and compare releases at the [GitHub release page](https://github.com/nuwave/lighthouse/releases).
 
 ## Unreleased
-- Added `@withCount` directive to allow eager loading relationship counts.
-- 
+
 ### Added
 
 - Write definitions for programmatically registered types to `programmatic-types.graphql`
   when running the `lighthouse:ide-helper` artisan command https://github.com/nuwave/lighthouse/pull/1371
+- Add `@withCount` directive to allow eager loading relationship counts https://github.com/nuwave/lighthouse/pull/1390
 
 ### Fixed